Fix `fold_handler` for when `result` is `nil`
Changes the logic in `fold_handler` to only iterate its `result`
argument if it is not nil.  Before this change, there some language
servers would return an error (`err ~= nil`), but no result (`result ==
nil`), which was not caught before attempting to iterate `result`.  This
change tweaks the error handling to catch both cases.